Added tracking for particles determined in .track file from Andrea Maccio. Indices start at 1 (not 0) to be consistent with tipsy index files.
Still one bad thing: can only read in xxx.track file where xxx = achInFile. That means at a restart one has also to rename the .track file. The problem is that in the main.c file the program reads first the tipsy file and then initialises the msr structure (and from there the pkd structure). I tried to swap the two lines in main.c but it did not work. Why? Is there an easy way to do it? An obvious way would be to write some functions..... Generally, it would be nice to have as filename achOutName => xxx = achOutName.

Marcel Zemp