Added a new analysis module MeasurePk which will measure the power spectrum of a given box using an FFT (FFTW).  The resulting P(k) still needs to be scaled by the volume of the box (pkdgrav does not know the volume).
Added support for compressed ASCII output.  If the filename for the ASCII
output ends in ".gz" then it is compressed using libz (gzip).  If the file
ends with ".bz2", then it is compressed using the bzip2 library.

Fixed some compilation problems caused when the prototypes for group finding
were changed.  This was only a problem when python was enabled.