A couple of experimental changes. First, there is now support for using CUDA to evaluate P-P interactions. This will only work with the null and pthread mdl at the moment (MPI has not been tested). This is enabled with:   ./configure --with-cuda=/usr/local/cuda
Next, AVX support has been added to the SIMD code. This should be enabled
automatically with "-march=native", but wasn't for some reason. The new
AVX code can be tested with:
  make pthreads CFLAGS="-g -O2 -mavx"

Both of these changes need to be fully regression tested, particularly
with regards to the GravStep time stepping criteria.

There were a number of other minor cleanups, most notably explicit casting
was added in many cases via the CAST macro. This allows the c++ compiler
to compile code (although this is not fully supported yet).