Added unbinding of groups created using the phase-space group finder.
A number of changes were required for this support. In particular,
multiple trees, one per group, were necessary to calculate gravity.
The routine that processes a checklist in walk2.c was split
preserve as much common code between the normal single tree walk
and the multiple group tree walk.

Additionally, the phase-space k-nearest neighbor code was moved
to knn6d.c and new routines for gather particles given a fixed
search radius was added. This improves the speed of the group
linking after the particles densities have been found.

The python code has preliminary support for copying parameter
files automatically to and from the python environment without
needing to explicitly copy values. This will need to be updated
to support undefined parameters and defaults.