Added correct support for unspecified parameters in a python script.
Parameters that are unspecified on the pkdgrav side are assigned a
value of None in the python script. If the value is still None when
the parameter list is updated, the parameter is left unspecified.

Removed a work around from master.c that forced the old domain
decomposition if iDomainMethod was specified and > -2. Leaving
iDomainMethod as None in a python script will cause the default
to be the old method.