Changed base SIMD Level to be SSE instead of SSE3 for SIMD_HADD()
SIMD_HADD() defaults to a regular loop if header __SSE3__ is not
defined because the _mm_hadd_ps() is not present in instrinic headers
for SSE and SSE2. In addition to adding support for SSE, _mm_hadd_ps()
calls were dropped in favour of lower latency and higher throughput
functions.

_mm_hadd_ps() has a latency of 5 clock cycles on archs older than
Skylake. In the event AVX is not used, the latency cost of 2 _mm_hadd_ps
calls have a total latency of 10 clock cycles, a lower latency solution
is presented using shuffle and duplication operations, which both have
latency and throughputs of 1 clock cycle.

Summary of changes:
- SIMD_HADD can now target SSE instead of SSE3
- SIMD_HADD still retains a target for SSE3
- Header guard added for <stdint.h> to avoid duplicate #include
- -O2 and -O3 for GCC produces a 2x faster performance in comparison to regular loop