Fixed scale-factor dependencies in hydroStep
Missing scale-factor dependencies were added to both
the CFL condition and the acceleration criteria.

With this, the overly conservative sum of the moduli
of the hydro+grav acceleration is no longer needed.

This should solve, or at least highly improve, the
"explosions" taking place in low density environments
for cosmological boxes.