PKDGRAV-1 Added lifetime function and inverse lifetime function.
Added functions to compute the numnber of SNIa events.

Added a check of the normalization of the IMF. The current implementation of the function stevChabrierIMF has been tested and found to give the correct normalization for any mass range.

Buffer arrays were transposed, they are now is ordered as (Elements, Metallicities, Masses) in row-major (right-most index changes the fastest).