PKDGRAV-1 Smooth functions have been written. The necessary helper functions to compute the ejected mass have been added. The distribution of this mass into neighboring gas particles is yet to be included.
The mass-interpolated data from the tables (what I call the buffer) has been transposed such that the arrangement is now (Metallicity, Mass, Element). In this way, it is easier to write general interpolation functions that work in 3D and 2D arrays, since now only the last axis is different.

Buffer metalicities are now in log. As is done in EAGLE, a logarithmic floor is used that is set to -20.

Parameters dCCSN_MinMass and dSNIa_MaxMass changed from double to float.

Removed unnecessary conversions to code units for some parameters.