PKDGRAV-1 The stellar evolution module has now been verified to give "reasonable" results for the ejecta. Verification is pending for its distribution to the surrounding gas particles.
STARFIELDS has been modified. It no longer stores the star particle's remaining
mass in each tracked element and metals. This was found to be meaningless since
this masses not only change due to ejection, but due to nucleosynthesis as well.
Instead, the cummulative ejected mass for each element and metals is now stored.
The necessary modifications in different parts of the code to account for this
change were carried out.

The inline stevGetIndex1DReversed has been added to take care of cases where
interpolation of an array in decreasing order is needed (e.g., inverse lifetime
function).

IMF-weighted integration of the ejecta is now carried out in log space. For this,
an array was added to inStellarEvolution (STEV_DATA) structure to store the
logarithms of the initial mass array.

Other minor changes to hopefully improve readability and to fix some bugs
identified during testing.