PKDGRAV-13 Improved class_get to clarify when 'classes' are used
There are three options:
a) Both 'classes' and 'class' fields are given and are used to
get the mass and softening
b) Only 'classes' is present, and the class of each particle is
inferred from the particle ordering
c) Neither 'classes' nor 'class' are present, and the softening
and masses are set using the particle array.

For options a) and b), the mass or softening can still be overriden
if there is a particle field for those.