PKDGRAV-18 Minor improvements to the stellar evolution module
Replaced log and pow functions with their single-precision version, but only in parts of the module that are not part of the initialization of the simulation.

Modified the integration of ejecta. The previous version was very sensitive to round-off errors.

Removed afLogMasses and afIMF arrays from the STEV_DATA structure. Added a float (fDeltaLogMass) and an array (afIMFLogWeights) in their replacement.

Improved the reading of yield tables to, hopefully, simplify making changes when datasets have different names and identify incorrect file setup.