PKDGRAV-82 Distribution of stev ejecta was divided into two steps
A first step accumulates mass, momentum and total energy in the new
variables pSph->[afReceivedMom,fReceivedMass,fReceivedE], and a second
step stores them and computes the associated change in thermal energy.

Furthermore, the site in the code where parameter bChemEnrich is evaluated
was moved, although its effect remains the same.