PKDGRAV-100 Added extra barrier before closing the caches
In some cases (sadly I was not able to isolate them, because
running the simulation again did not crash), the caches try
to do some work after being closed, leading to a segfault.

I do not see a reason for this to happen. But adding an extra
barrier seems to help, with little to zero extra overhead.