Fixed a bug in setting the gas particle position when adding baryons to the ICs generated by pkdgrav3.
Gas particles were displaced by half cell size along each
coordinate. This does not conserve the centre of mass of the original
particle, shifts non-uniformly P(k) phases, and causes a drift of the
box towards the direction of the displacement.

The new implementation does keep the gas particles on top of the dark
matter particles, as long as nBucket > 1.

If nBucket = 1, both gas and dark matter particles are displaced in
opposite direction and proportionally to their mass, as in N-GenIC and
as originally meant to be. This conserves the centre of mass, linear
and angular momentum, and kinetic energy. Particles are displaced by a
quantity proportional to 1/25 the cell size, about the value of the
softening length, in order to minimise inaccuracies.