Update of the kernel function: we define the cubic spline kernel as function of r/H with H the compact support radius. The smoothing length is then h = H/2. This way there is no need to compute h in most of the code that makes use of the kernel. The value of h is used in a couple of places where the "spatial resolution" is needed for the computation.
Fixed the computation of the Courant hydro timestep. There, the timestep
was defined as:

dt = C * 2*h/v_sig

This was probably taken from the definition in Gadget:

dt = C * H/v_sig

therefore the 2*h. However, Gadget uses a value of C that is half what
normally suggested, to compensate for the use of H in the formula.

Following Monaghan original derivation,

dt = C h/v_sig

with C~0.3-0.4, thus the value we are using.