Unify simd versions of the hydrodynamic solver
There is no code duplication!

A class was created for the Riemann solver and
another for the meshless solver. They are templated
by the vector type and the mask.
They have been tested to work for the architecture-defined
`dvec` and the no-simd fallback `vec<double,double>`.

NOTE: The MFV simd implementation is working but not optimised.